requirements.txt
pandas
numpy
yfinance
scikit-learn
statsmodels
matplotlib
pyyaml
pytest
colorama
streamlit
plotly
